projet de mise en situation/projet de mise en situation/backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
sqlmodel>=0.0.22,<0.0.23
pydantic>=2.8,<3.0
python-multipart>=0.0.9,<0.1
passlib[bcrypt]>=1.7,<1.8
python-jose[cryptography]>=3.3,<3.4
python-dotenv>=1.0,<2.0
PyMuPDF>=1.24,<1.25
python-docx>=1.1,<2.0
chardet>=5.2,<6.0
scikit-learn>=1.4,<2.0
numpy>=1.26,<3.0
sentence-transformers>=2.2,<3.0

email-validator>=2,<3
